Moving intersection method from DatetimeIndex and Timedeltaindex to datetimelike. Point PeriodIndex intersection to Index.intersection
